fix: make check breaking due to cargo deny breaking changes
Due to cargo deny changes that were mentioned in #89, I did some
changes to `deny.toml` in order to fix it.

Although in the issue I mentioned the idea of using `cd-allow` to add
all ofi and fsf licenses into `deny.toml` to keep the same
functionality, when implementing I saw that some licenses listed in
[SPDX](https://spdx.org/licenses) are not recognized by cargo deny,
for example `AFL-1.1`. So as this repo and its dependencies are all
below `Apache-2.0` license (at least according to cargo deny), I think
it's best to just add it to the configuration file and whenever this
change it tackles the problem